+++ cmd/ergo/filelock_unix_test.go
@@ -11,6 +11,10 @@ func lockTestFile(file *os.File) error {
 	return syscall.Flock(int(file.Fd()), syscall.LOCK_EX|syscall.LOCK_NB)
 }
 
+func lockTestFileShared(file *os.File) error {
+	return syscall.Flock(int(file.Fd()), syscall.LOCK_SH|syscall.LOCK_NB)
+}
+
 func unlockTestFile(file *os.File) error {
 	return syscall.Flock(int(file.Fd()), syscall.LOCK_UN)
 }