+++ cmd/ergo/filelock_windows_test.go
@@ -20,6 +20,18 @@ func lockTestFile(file *os.File) error {
 	)
 }
 
+func lockTestFileShared(file *os.File) error {
+	var overlapped windows.Overlapped
+	return windows.LockFileEx(
+		windows.Handle(file.Fd()),
+		windows.LOCKFILE_FAIL_IMMEDIATELY,
+		0,
+		1,
+		0,
+		&overlapped,
+	)
+}
+
 func unlockTestFile(file *os.File) error {
 	var overlapped windows.Overlapped
 	return windows.UnlockFileEx(windows.Handle(file.Fd()), 0, 1, 0, &overlapped)